Set Prometheus retention to 30d
Allows us to keep Sloth recording rules that use `[30d]` on Prometheus
and evaluate currectly.

Also expose a flag to set retention size, we can set this just under the
size of the volume so in case the volume is too small we drop data
instead of filling it and crashing.

Remove `block-duration` flags - they don't exist in docs for the latest
version:
https://prometheus.io/docs/prometheus/2.54/command-line/prometheus/ - so
I doubt they have any effect.